+++ src/types.ts
@@ -17,6 +17,7 @@ export type CardConfig = {
   hide_arm_options: boolean;
   pending_sound: string;
   click_sound: string;
+  disarm_sound: string;
 };
 
 export type HideOptions = 'always' | 'never' | 'disarmed' | 'armed';